+++ vlib/crypto/sha256/sha256block_generic.v
@@ -13,7 +13,7 @@ import math.bits
 
 const (
 	_K = [
-		0x428a2f98,
+		u32(0x428a2f98),
 		0x71374491,
 		0xb5c0fbcf,
 		0xe9b5dba5,
